requirements.txt
Flask>=3.1.0

pytesseract>=0.3.13
pillow>=11.1.0
Werkzeug>=3.1.3
